requirements.txt
asgiref==3.7.2
Django==4.2.3
django-debug-toolbar==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
